gdpc features added
