+++ grimoire/core/structures/nbt/build_nbt.py
@@ -2,6 +2,7 @@ from typing import Callable
 
 from gdpc import Block
 from gdpc.editor import Editor
+from gdpc.lookup import TRAPDOORS, DOORS, FENCES, WALLS, SLABS, STAIRS
 from gdpc.vector_tools import ivec3
 
 from grimoire.core.styling.legacy_palette import LegacyPalette, palette_swap
@@ -72,7 +73,8 @@ def build_nbt(
 
         # Doesn't allow non-solid blocks to replace blocks
         if (not allow_non_solid_replacement) and any(
-            blocktype in block.id for blocktype in ("stairs", "slab", "walls", "fence")
+            blocktype in block.id
+            for blocktype in STAIRS | SLABS | WALLS | FENCES | DOORS | TRAPDOORS
         ):
             curr_block = editor.getBlock(position=(x, y, z))
